modernize - some additional refactoring
