+++ index.js
@@ -6,5 +6,7 @@ module.exports = {
     concurrent: require('./lib/concurrent.js').concurrent,
     concurrentAll: require('./lib/concurrent-all.js').concurrentAll,
     mapConcurrent: require('./lib/map-concurrent.js').mapConcurrent,
-    mapConcurrentAll: require('./lib/map-concurrent-all.js').mapConcurrentAll
+    mapConcurrentAll: require('./lib/map-concurrent-all.js').mapConcurrentAll,
+    throttledConcurrentAll: require('./lib/throttled-concurrent-all.js').throttledConcurrentAll,
+    throttledMapConcurrentAll: require('./lib/throttled-map-concurrent-all.js').throttledMapConcurrentAll
 };
