+++ lib/concurrent-all.js
@@ -8,9 +8,7 @@ module.exports = {
 
         const doConcurrently = (fn, index, state, done) => {
             fn((error, result) => {
-                if (error) {
-                    state.numErrors = state.numErrors ? state.numErrors + 1 : 1;
-                }
+                if (error) ++state.numErrors;
 
                 state.results[index] = {
                     error: error || null,
@@ -27,10 +25,9 @@ module.exports = {
 
         const initialState = {
             numToDo: functions.length,
-            numErrors: null,
+            numErrors: 0,
             numComplete: 0,
-            results: [],
-            error: null
+            results: []
         };
 
         functions.forEach((fn, index) => {
