+++ lib/map-concurrent.js
@@ -4,11 +4,7 @@ const concurrent = require('./concurrent.js').concurrent;
 
 module.exports = {
     mapConcurrent: function(items, iteratee, callback) {
-
-        const toRun = items.map(function(item) {
-            return iteratee.bind(null, item);
-        });
-
+        const toRun = items.map((item) => iteratee.bind(null, item));
         concurrent(toRun, callback);
     }
 };
