+++ lib/throttled-concurrent-all.js
@@ -1,26 +1,18 @@
 'use strict';
 
 module.exports = {
-    throttledConcurrentAll(functionsToRun, numWorkers, allFunctionsCalled) {
+    throttledConcurrentAll(functionsToRun, numWorkers, callback) {
 
         if (typeof (numWorkers) === 'function') {
-            allFunctionsCalled = numWorkers;
+            callback = numWorkers;
             numWorkers = Math.floor(functionsToRun.length / 2);
         }
 
-        if (functionsToRun.length === 0) { return allFunctionsCalled(null, []); }
+        if (functionsToRun.length === 0) { return callback(null, []); }
 
         const actualNumWorkers = functionsToRun.length < numWorkers ? functionsToRun.length : numWorkers;
 
-        const state = {
-            functionsToRun: functionsToRun,
-            results: [],
-            indexToProcess: 0,
-            numComplete: 0,
-            numErrors: 0
-        };
-
-        function worker() {
+        const concurrentWorker = (state, done) => {
 
             if (state.numComplete === state.functionsToRun.length) { return; }
 
@@ -34,22 +26,32 @@ module.exports = {
 
             toDo(function(error, result) {
                 if (error) ++state.numErrors;
+
                 state.results[resultPosition] = {
-                    error: error,
+                    error: error || null,
                     result: error ? null : result
                 };
+
                 ++state.numComplete;
 
                 if (state.numComplete === state.functionsToRun.length) {
-                    allFunctionsCalled(state.numErrors, state.results);
-                } else {
-                    worker();
+                    return done(state.numErrors || null, state.results);
                 }
+
+                process.nextTick(() => concurrentWorker(state, done));
             });
-        }
+        };
+
+        const initialState = {
+            functionsToRun: functionsToRun,
+            results: [],
+            indexToProcess: 0,
+            numComplete: 0,
+            numErrors: 0
+        };
 
         for (let x = 0; x < actualNumWorkers; ++x) {
-            worker();
+            process.nextTick(() => concurrentWorker(initialState, callback));
         }
     }
 };
