+++ lib/throttled-map-concurrent-all.js.js
@@ -1,26 +1,18 @@
 'use strict';
 
 module.exports = {
-    throttledMapConcurrentAll(items, iteratee, numWorkers, allFunctionsCalled) {
+    throttledMapConcurrentAll(items, iteratee, numWorkers, callback) {
 
         if (typeof (numWorkers) === 'function') {
-            allFunctionsCalled = numWorkers;
+            callback = numWorkers;
             numWorkers = Math.floor(items.length / 2);
         }
 
-        if (items.length === 0) { return allFunctionsCalled(null, []); }
+        if (items.length === 0) { return callback(null, []); }
 
         const actualNumWorkers = items.length < numWorkers ? items.length : numWorkers;
 
-        const state = {
-            itemsToProcess: items,
-            results: [],
-            indexToProcess: 0,
-            numComplete: 0,
-            numErrors: 0
-        };
-
-        function worker() {
+        function concurrentWorker(state, done) {
 
             if (state.numComplete === state.itemsToProcess.length) { return; }
 
@@ -34,22 +26,33 @@ module.exports = {
 
             iteratee(itemToProcess, function(error, result) {
                 if (error) ++state.numErrors;
+
                 state.results[resultPosition] = {
-                    error: error,
+                    error: error || null,
                     result: error ? null : result
                 };
+
                 ++state.numComplete;
 
                 if (state.numComplete === state.itemsToProcess.length) {
-                    allFunctionsCalled(state.numErrors, state.results);
-                } else {
-                    worker();
+                    return done(state.numErrors || null, state.results);
                 }
+
+                process.nextTick(() => concurrentWorker(state, done));
             });
         }
 
+        const state = {
+            itemsToProcess: items,
+            iteratee: iteratee,
+            results: [],
+            indexToProcess: 0,
+            numComplete: 0,
+            numErrors: 0
+        };
+
         for (let x = 0; x < actualNumWorkers; ++x) {
-            worker();
+            concurrentWorker(state, callback);
         }
     }
 };
