+++ test/unit/index-tests.js
@@ -13,84 +13,64 @@ describe('index - unit tests', function() {
     let concurrentAll;
     let mapConcurrent;
     let mapConcurrentAll;
+    let throttledMapConcurrentAll;
+    let throttledConcurrentAll;
     let palinode;
 
-    before(function() {
+    beforeEach(function() {
         mockery.enable({useCleanCache: true});
         mockery.registerAllowable('../../index.js');
-    });
 
-    after(function() {
-        mockery.deregisterAll();
-        mockery.disable();
-    });
+        mockery.registerMock('./lib/series.js', {series: series = sinon.stub()});
+        mockery.registerMock('./lib/map-each.js', {mapEach: mapEach = sinon.stub()});
+        mockery.registerMock('./lib/concurrent.js', {concurrent: concurrent = sinon.stub()});
+        mockery.registerMock('./lib/concurrent-all.js', {concurrentAll: concurrentAll = sinon.stub()});
+        mockery.registerMock('./lib/map-concurrent.js', {mapConcurrent: mapConcurrent = sinon.stub()});
+        mockery.registerMock('./lib/map-concurrent-all.js', {mapConcurrentAll: mapConcurrentAll = sinon.stub()});
+        mockery.registerMock('./lib/throttled-map-concurrent-all.js', {throttledMapConcurrentAll: throttledMapConcurrentAll = sinon.stub()});
+        mockery.registerMock('./lib/throttled-concurrent-all.js', {throttledConcurrentAll: throttledConcurrentAll = sinon.stub()});
 
-    beforeEach(function() {
-        series = sinon.stub();
-        mapEach = sinon.stub();
-        concurrent = sinon.stub();
-        concurrentAll = sinon.stub();
-        mapConcurrent = sinon.stub();
-        mapConcurrentAll = sinon.stub();
-
-        mockery.registerMock('./lib/series.js', {series: series});
-        mockery.registerMock('./lib/map-each.js', {mapEach: mapEach});
-        mockery.registerMock('./lib/concurrent.js', {concurrent: concurrent});
-        mockery.registerMock('./lib/concurrent-all.js', {concurrentAll: concurrentAll});
-        mockery.registerMock('./lib/map-concurrent.js', {mapConcurrent: mapConcurrent});
-        mockery.registerMock('./lib/map-concurrent-all.js', {mapConcurrentAll: mapConcurrentAll});
-
+        palinode = require('../../index.js');
     });
 
     afterEach(function() {
-        mockery.resetCache();
-        mockery.deregisterMock('./lib/series.js');
-        mockery.deregisterMock('./lib/map-each.js');
-        mockery.deregisterMock('./lib/concurrent.js');
-        mockery.deregisterMock('./lib/concurrent-all.js');
-        mockery.deregisterMock('./lib/map-concurrent.js');
-        mockery.deregisterMock('./lib/map-concurrent-all.js');
+        mockery.deregisterAll();ß
+        mockery.disable();
     });
 
     it('should have 6 properties', function() {
-        palinode = require('../../index.js');
-
-        expect(Object.keys(palinode).length).to.equal(6);
+        expect(Object.keys(palinode).length).to.equal(8);
     });
 
     it('should have the property \'series\' with the value from require(\'./lib/series.js\').series', function() {
-        palinode = require('../../index.js');
-
         expect(palinode.series).to.deep.equal(series);
     });
 
     it('should have the property \'mapEach\' with the value from require(\'./lib/map-each.js\').mapEach', function() {
-        palinode = require('../../index.js');
-
         expect(palinode.mapEach).to.deep.equal(mapEach);
     });
 
     it('should have the property \'concurrent\' with the value from require(\'./lib/concurrent.js\').concurrent', function() {
-        palinode = require('../../index.js');
-
         expect(palinode.concurrent).to.deep.equal(concurrent);
     });
 
     it('should have the property \'concurrentAll\' with the value from require(\'./lib/concurrent-all.js\').concurrentAll', function() {
-        palinode = require('../../index.js');
-
         expect(palinode.concurrentAll).to.deep.equal(concurrentAll);
     });
 
     it('should have the property \'mapConcurrent\' with the value from require(\'./lib/map-concurrent.js\').mapConcurrent', function() {
-        palinode = require('../../index.js');
-
         expect(palinode.mapConcurrent).to.deep.equal(mapConcurrent);
     });
 
     it('should have the property \'mapConcurrentAll\' with the value from require(\'./lib/map-concurrent-all.js\').mapConcurrentAll', function() {
-        palinode = require('../../index.js');
-
         expect(palinode.mapConcurrentAll).to.deep.equal(mapConcurrentAll);
+    });
+
+    it('should have the property \'throttledMapConcurrentAll\' with the value from require(\'./lib/throttled-map-concurrent-all.js\').throttledMapConcurrentAll', function() {
+        expect(palinode.throttledMapConcurrentAll).to.deep.equal(throttledMapConcurrentAll);
+    });
+
+    it('should have the property \'throttledConcurrentAll\' with the value from require(\'./lib/throttled-concurrent-all.js\').throttledConcurrentAll', function() {
+        expect(palinode.throttledConcurrentAll).to.deep.equal(throttledConcurrentAll);
     });
 });
